requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-apscheduler==0.6.2
django-filter==23.2
flake8==6.0.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.5
requests==2.29.0
requests-oauthlib==1.3.1
shortcuts==0.11.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
